# docs/handoff/ups-expresosapp-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/stops": {
      "post": {
        "operationId": "StopsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStopDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stop created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new stop",
        "tags": [
          "Admin Stops"
        ]
      },
      "get": {
        "operationId": "StopsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "description": "Page number (1-based)",
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "description": "Items per page",
            "required": false,
            "name": "limit",
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of stops",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopPaginatedResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all stops with pagination",
        "tags": [
          "Admin Stops"
        ]
      }
    },
    "/admin/stops/{id}": {
      "get": {
        "operationId": "StopsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Stop ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stop details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Stop not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get stop details",
        "tags": [
          "Admin Stops"
        ]
      },
      "patch": {
        "operationId": "StopsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Stop ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStopDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stop updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Stop not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a stop",
        "tags": [
          "Admin Stops"
        ]
      }
    },
...
    "/trip-feedback": {
      "post": {
        "operationId": "TripFeedbackController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripFeedbackResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Route or driver not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create trip feedback",
        "tags": [
          "Trip Feedback"
        ]
      },
      "get": {
        "operationId": "TripFeedbackController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by user ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "routeId",
            "required": false,
            "in": "query",
            "description": "Filter by route ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of feedback",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripFeedbackPaginatedResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List trip feedback with pagination",
        "tags": [
          "Trip Feedback"
        ]
      }
    },
    "/trip-feedback/{id}": {
      "get": {
        "operationId": "TripFeedbackController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Feedback ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripFeedbackResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Feedback not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get trip feedback by ID",
        "tags": [
          "Trip Feedback"
        ]
      }
    }
  },
  "info": {
    "title": "UPS ExpresosApp API",
    "description": "Backend API for UPS ExpresosApp MVP - Institutional transport management system for Universidad Politecnica Salesiana.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateStopDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Stop name",
            "example": "Parque de la Madre"
          },
          "reference": {
            "type": "string",
            "description": "Reference or address",
            "example": "Av. 12 de Abril y Loja"
          },
          "latitude": {
            "type": "number",
            "description": "Latitude coordinate",
            "minimum": -90,
            "maximum": 90,
            "example": -2.8975
          },
          "longitude": {
            "type": "number",
            "description": "Longitude coordinate",
            "minimum": -180,
            "maximum": 180,
            "example": -79.0045
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the stop is active",
            "default": true
          }
        },
        "required": [
          "name",
          "latitude",
          "longitude"
        ]
      },
      "StopResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "example": "Parque de la Madre"
          },
          "reference": {
            "type": "object",
            "example": "Av. 12 de Abril y Loja",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "example": -2.8975
          },
          "longitude": {
            "type": "number",
            "example": -79.0045
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "reference",
          "latitude",
          "longitude",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "StopPaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StopResponseDto"
            }
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateStopDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Stop name",
            "example": "Parque de la Madre"
          },
          "reference": {
            "type": "string",
            "description": "Reference or address",
            "example": "Av. 12 de Abril y Loja"
          },
          "latitude": {
            "type": "number",
            "description": "Latitude coordinate",
            "minimum": -90,
            "maximum": 90,
            "example": -2.8975
          },
          "longitude": {
            "type": "number",
            "description": "Longitude coordinate",
            "minimum": -180,
            "maximum": 180,
            "example": -79.0045
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the stop is active",
            "default": true
          }
        }
      },
...
      "CreateTripFeedbackDto": {
        "type": "object",
        "properties": {
          "routeId": {
            "type": "string",
            "description": "ID of the route being reviewed",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "driverId": {
            "type": "string",
            "description": "ID of the driver (optional)",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "rating": {
            "type": "number",
            "description": "Rating from 1 to 5",
            "example": 4,
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "description": "Optional comment",
            "example": "El servicio fue puntual y comodo"
          },
          "travelDate": {
            "type": "string",
            "description": "Date of travel (ISO 8601)",
            "example": "2026-07-01T08:00:00.000Z"
          }
        },
        "required": [
          "routeId",
          "rating"
        ]
      },
      "TripFeedbackResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "userId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "routeId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440002"
          },
          "driverId": {
            "type": "object",
            "example": "550e8400-e29b-41d4-a716-446655440003",
            "nullable": true
          },
          "rating": {
            "type": "number",
            "example": 4
          },
          "comment": {
            "type": "object",
            "example": "El servicio fue puntual",
            "nullable": true
          },
          "travelDate": {
            "type": "object",
            "example": "2026-07-01T08:00:00.000Z",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-07-01T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "routeId",
          "rating",
          "createdAt"
        ]
      },
      "TripFeedbackPaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripFeedbackResponseDto"
            }
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      }
    }
  }
}